Changed Cache logic namespace
